+++ lib/petitest/test_groups.rb
@@ -1,0 +1,4 @@
+module Petitest
+  module TestGroups
+  end
+end